pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vizyonai"
version = "0.1.0"
description = "VizyonAI retail assistant"
requires-python = ">=3.10"
dependencies = [
  "streamlit",
  "python-dotenv",
  "openai",
  "pandas",
  "rapidfuzz"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
